Decompose engine.ts into gates, subgraph, state, and transitions modules
engine.ts was 664 lines with advance() as a 170-line god method handling
wait blocking, return schema, validations, edge conditions, subgraph
push/pop, and terminal detection inline. Split into focused modules:

- gates.ts: pre-advance checks (wait, returns, validations, edge conditions)
- subgraph.ts: stack push/pop with context and return mapping
- state.ts: context updates, strict context, inspect builders, turn warnings
- transitions.ts: edge evaluation with default-edge logic

Also fixes: exhaustive switch in inspect(), non-null assertions replaced
with defensive checks, and duplicated error construction consolidated
into makeAdvanceError helper.
