+++ src/engine/subgraph.ts
@@ -1,0 +1,157 @@
+import { EngineError } from "../errors.js";
+import { evaluate } from "../evaluator.js";
+import { cloneContext, toNodeInfo } from "./helpers.js";
+import { evaluateTransitions } from "./transitions.js";
+import type {
+  ValidatedGraph,
+  NodeDefinition,
+  SessionState,
+  AdvanceSuccessResult,
+} from "../types.js";
+
+export function maybePushSubgraph(
+  stack: SessionState[],
+  graphs: Map<string, ValidatedGraph>,
+  previousNode: string,
+  edge: string,
+  newNodeDef: NodeDefinition,
+  maxDepth: number
+): AdvanceSuccessResult {
+  const parentSession = stack[stack.length - 1];
+  const subgraph = newNodeDef.subgraph!;
+
+  if (subgraph.condition) {
+    let condMet: boolean;
+    try {
+      condMet = evaluate(subgraph.condition, parentSession.context);
+    } catch {
+      condMet = false;
+    }
+    if (!condMet) {
+      return {
+        status: "advanced",
+        isError: false,
+        previousNode,
+        edgeTaken: edge,
+        currentNode: parentSession.currentNode,
+        node: toNodeInfo(newNodeDef),
+        validTransitions: evaluateTransitions(newNodeDef, parentSession.context),
+        context: cloneContext(parentSession.context),
+      };
+    }
+  }
+
+  if (stack.length >= maxDepth) {
+    throw new EngineError(
+      `Maximum stack depth (${maxDepth}) exceeded. Cannot push subgraph '${subgraph.graphId}'. Simplify the workflow or increase maxDepth.`,
+      "STACK_DEPTH_EXCEEDED"
+    );
+  }
+
+  const childGraph = graphs.get(subgraph.graphId);
+  if (!childGraph) {
+    throw new EngineError(
+      `Subgraph '${subgraph.graphId}' not found in loaded graphs.`,
+      "GRAPH_NOT_FOUND"
+    );
+  }
+
+  const childDef = childGraph.definition;
+  const childContext: Record<string, unknown> = {
+    ...(childDef.context ?? {}),
+    ...(subgraph.initialContext ?? {}),
+  };
+
+  if (subgraph.contextMap) {
+    for (const [parentKey, childKey] of Object.entries(subgraph.contextMap)) {
+      if (parentKey in parentSession.context) {
+        childContext[childKey] = parentSession.context[parentKey];
+      }
+    }
+  }
+
+  stack.push({
+    graphId: subgraph.graphId,
+    currentNode: childDef.startNode,
+    context: childContext,
+    history: [],
+    contextHistory: [],
+    turnCount: 0,
+    startedAt: new Date().toISOString(),
+  });
+
+  const childStartNode = childDef.nodes[childDef.startNode];
+  const activeSession = stack[stack.length - 1];
+
+  return {
+    status: "advanced",
+    isError: false,
+    previousNode,
+    edgeTaken: edge,
+    currentNode: childDef.startNode,
+    subgraphPushed: {
+      graphId: subgraph.graphId,
+      startNode: childDef.startNode,
+      stackDepth: stack.length,
+    },
+    node: toNodeInfo(childStartNode),
+    validTransitions: evaluateTransitions(childStartNode, activeSession.context),
+    context: cloneContext(activeSession.context),
+  };
+}
+
+export function popSubgraph(
+  stack: SessionState[],
+  graphs: Map<string, ValidatedGraph>,
+  previousNode: string,
+  edge: string
+): AdvanceSuccessResult {
+  const childSession = stack[stack.length - 1];
+  const completedGraphId = childSession.graphId;
+
+  stack.pop();
+
+  const parentSession = stack[stack.length - 1];
+  const parentGraph = graphs.get(parentSession.graphId);
+  if (!parentGraph) {
+    throw new EngineError(
+      `Graph "${parentSession.graphId}" not found`,
+      "GRAPH_NOT_FOUND"
+    );
+  }
+  const parentDef = parentGraph.definition;
+  const parentNodeDef = parentDef.nodes[parentSession.currentNode];
+  const subgraphDef = parentNodeDef.subgraph!;
+
+  const returnedContext: Record<string, unknown> = {};
+  if (subgraphDef.returnMap) {
+    for (const [childKey, parentKey] of Object.entries(subgraphDef.returnMap)) {
+      if (childKey in childSession.context) {
+        const value = childSession.context[childKey];
+        parentSession.context[parentKey] = value;
+        returnedContext[parentKey] = value;
+        parentSession.contextHistory.push({
+          key: parentKey,
+          value,
+          setAt: parentSession.currentNode,
+          timestamp: new Date().toISOString(),
+        });
+      }
+    }
+  }
+
+  return {
+    status: "subgraph_complete",
+    isError: false,
+    previousNode,
+    edgeTaken: edge,
+    currentNode: parentSession.currentNode,
+    completedGraph: completedGraphId,
+    returnedContext,
+    stackDepth: stack.length,
+    resumedNode: parentSession.currentNode,
+    node: toNodeInfo(parentNodeDef),
+    validTransitions: evaluateTransitions(parentNodeDef, parentSession.context),
+    context: cloneContext(parentSession.context),
+  };
+}
